Cleanups and working save/load functionality
For now, this will only load blocks, not heaps. This means we must
re-compute all blocks when relaunching the node process. Heaps are
already saved, but not loaded. Loading them will massively speed up the
time it takes to open the node, but we need more tests to do that
confidently.