U120, I120, UTUP, ITUP operations WIP
Kindelia will have 4 numeric types:

- U120: a 120-bit unsigned integer
- I120: a 120-bit integer
- UTUP: a 8-bit, a 16-bit, a 32-bit and a 64-bit unsigned integer
- ITUP: a 8-bit, a 16-bit, a 32-bit and a 64-bit integer

The TUP types store the 4 primitive integers unboxedly in a single HVM
link. Note that 8 + 16 + 32 + 64 = 120, so the entire space is used.

This commit implements UTUP only. UTUP operations can be written with a
tilde (~) followed by the usual operation symbol. For example:

(~+ a b)

This will perform an addition on the elements of the 'a' and 'b' utups.
This addition will be performed element-wise, i.e., the 8-bit uint on
'a' will be added to the 8-bit uint on 'b', and so on.

This commit also implements rotate_left and rotate_right operations,
written as (~<~ a b) and (~>~ a b), for the UTUP type. This fixes #35,
making SHA3 considerably more efficient.