Initial HTTP API setup
Now, accessing http://localhost:8000/get_tick will show the current tick
on the node process. The HTTP API is spawned on a separate thread, and
oneshot channels are used to communicate with the main node thread. The
"Request" type on node.rs will have all the possible methods. The rest
of the external API will be designed that way.