ECDSA account system. Sorry.
Kindelia now uses the same accounts as Ethereum. Run statements can now be signed:

run {
  code_here
} sign {
  signature_here
}

The payload is the serialization of the block. If there is no signature, or if
ECDSA.verify fails, the statement will run with the subject set to "0". If it
checks, the statement runs as usual, except the subject, instead of "0", will
store the last 120 bits of the corresponding Ethereum address. This allows
functions to address users by u120 (20-letter) names.

At this point, I can already hear Kelvin yelling:

"Why... why would you, betray us all like that? How dare you ruin this
magnificently pure, strong network with such a complex, convoluted, fragile
cryptographic primitive that is about to be broken into pieces, either by the
rise of quantum computers, or even classical solutions? We haven't even had a
chance to talk this over! Can't you think of the future children? Can't you
think of the future cats? Why would you put so much effort into designing a
network capable of lasting centuries... only to ruin it overnight?

For him, for future users, for myself, I can only admit: no network can last a
hundred years if it never launches, to begin with. Truth is, it is 2022, we
won't get any kind of reasonable funding if we don't present a working network
with great apps that people can use to buy things and do stuff, today. Lamport
signatures are so huge they don't fit in an entire block, WOTS is extremelly
inefficient, implementing ECDSA on HVM will take months, which we don't have.
All that to end up with inefficient, error-prone, roll-your-own-crypto account
systems on HVM. And until then, there won't be many exciting apps running on
Kindelia. Perhaps anonymous boards, an /r/place clone, things like that, but not
anything that requires authentication. We can avoid all this pain with one line:

  external crate secp256k1;

Sounds tempting, doesn't it? We aren't in a bargaining position. If it helps,
the network will NOT be destroyed by a sudden rise of Quantum computers, unlike
Bitcoin and Ethereum. After all, we don't *rely* on ECDSA. We just use it on the
short term. We can simply demand users to assign hash signatures for backup,
allowing us to just disable ECDSA and let the network recover after such event.
All in all, if Kindelia's most unjustified, fragile aspect is the use of the
most reviewed signature standard in existence, which is also used by every other
project, I guess we're in good shoes.