Transaction posting and mempool functionality. Refactors.
- Now the deserialize functions return an Option

- `kindelia start` doesn't ask a file anymore

- Transactions can be posted with `kindelia post tx_file.kdl target_node_ip`

Posted transactions are added to the node's mempool. A node will build
the body of the blocks it attempts to mine by taking the first non-mined
transactions on the mempool, in order of hash (higher hash = higher
priority). The mempool is NOT propagated yet, so, if you want multiple
nodes to attempt to mine your transaction, you must send it to all these
nodes manually.