Fix incorrect behavior on default cases
Ported from an HVM fix. Below is its commit message:

Previously, the following program:

(Foo True) = True
(Foo a)    = False
Main       = @a (Foo a)

Would evaluate to `False`, which is incorrect, since the first argument
of `Foo` pattern-matches, thus, the variable should cause the call to
`Foo` on `Main` to be stuck.