Fixing U120 operations and divison by 0
Fixing the semantics of operations on u120 by adding special functions
to handle them.

Addition is semantically the same, ie. add as u128 and
then chop off at 120 bits.

Subtraction was changed to being addition with the 2's-complement
(which i don't think adds enough overhead to care), because messing
with the carry bit is messy.

Multiplication was changed to being long-form multiplication, which
again I don't think generates that much additional overhead. Honestly
I'm not sure if this is the best way to do it, but it works and it is
safe.

Division is exactly the same as u128 division, because if both numbers
fit in a u120, their quocient also will.

Remainder is also the same.

Shift left and shift right are a bit tricky, because shifting for more
bits than the type's width is generally undefined behaviour. Because
of that, it is defined that the second operand is taken modulo the
bit-size (u120).

Also added the check to division by 0, and throws DivisionByZero
error whenever it would panic.