Cell, RawCell and Loc types
the `Cell` type represents a KHVM term low level representation. it is
mostly used to expose the internals of `RawCell`.

`RawCell` is the numeric representation of a `Cell` as a raw
pointer (a `u128`). it is used to add type safety and cohesion to all
the usages of pointers inside the code.

the `Loc` type represents the u48 pointer inside a `RawCell` and is
used to add safety and semantics to the types that each function returns.