Allows `fun` deficnition to have optional initial state
Previous behavior was that functions with no declared initial
state actually had an default initial state of `#0`. That was
inconsistent with the behavior of IO `TAKE`, which leaves
 functions with no state at all.

Function state will be initialized by default to the "hole" cell
value `U128_NONE` when initial state is not present.

`TAKE` IO will also set function state to `U128_NONE`.
That is ok because function state holes on Heap layers are stored
on the maps as `Option`s so the actual `Some` value being
`U128_NONE` can represent a missing state on a written layer.

`fun` serializaition was changed to preceed initial state serialization
with one bit indicating presence. That should save some bits of block
space as most functions will be pure and so won't have initial state.