change If matching
instead of matching anything in the true case, match specifically only
the number 1.

otherwise it will always match weird terms, for example the following
term `(== #5 {Ctr #5})` does not reduce and it will still evaluate to
true, which is at least unexpected.

maybe introduce {True} and {False} in order to make this a little bit
more "type safe"?